Split and organize the codes
